Isolate lyrics feature in a separate module
