+++ bon-jova-quarkus-extension/lyrics/src/test/java/org/example/bon/jova/quarkus/extension/lyrics/LyricsReaderTest.java
@@ -1,5 +1,6 @@
-package org.example.bon.jova.quarkus.extension.deployment.lyrics;
+package org.example.bon.jova.quarkus.extension.lyrics;
 
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
@@ -11,14 +12,14 @@ class LyricsReaderTest {
     @Test
     void readLyricsShouldReturn100Songs() {
         List<String> allLyrics = LyricsReader.readAll(true);
-        assertEquals(100, allLyrics.size());
+        Assertions.assertEquals(100, allLyrics.size());
     }
 
     @Test
     void readLyricsTwiceShouldUseTheCache() {
         List<String> allLyrics = LyricsReader.readAll(true);
         List<String> allLyrics2 = LyricsReader.readAll(true);
-        assertEquals(100, allLyrics.size());
-        assertEquals(100, allLyrics2.size());
+        Assertions.assertEquals(100, allLyrics.size());
+        Assertions.assertEquals(100, allLyrics2.size());
     }
 }
