+++ bon-jova-quarkus-extension/pom.xml
@@ -17,8 +17,9 @@
 
     <modules>
         <module>deployment</module>
-        <module>runtime</module>
         <module>integration-tests</module>
+        <module>runtime</module>
+        <module>lyrics</module>
     </modules>
 
     <properties>
